More rule stack work
This commit contains most of the work to get rule stacks
implemented. It replaces Flatten with a new trait called Execute,
which operates on a &mut Grammar so that the Grammar is the single
source of truth for keys and rules. It supports pushing new rules onto
the stack for a given key, but does not yet support popping.

This commit also sets the stage for later work where having the
Grammar maintain state between executions will be useful.